WIP: Versioned index home page for ATK docs (BROKEN)
Problems:

1. The CSS is messed up so that the dropdown menu doesn't currently
display when you roll over the main menu, so you can't actually click
on anything.  Oops.

2. The CSS is messed up so that the menus don't appear in the right
place; the top menu that says "Version 1.x" right now is supposed to
be inside the "header" and "header-container" `div`s, at the bottom
right of the blue area, with the right edge against the right margin
of the text in the "content" `div`.

3. I thought I was being really clever by loading the index page with
AJAX into a `div` so it scrolls under an absolute-positioned header,
but I'm not sure what's going to happen when someone clicks on a link
inside that text.  Won't it just take them to a regular old HTML page
without the header?  Then they lose the version menu, and the whole
point was to keep the version menu and the header in place regardless
of what's being shown inside the "content" `div`.

If this is the case then the only way I can think of to fix this right
now is to add javascript to intercept all click events on `a` tags and
replace each one with another AJAX call like the first one, as long as
it points to a file on the local server, so that the new page loads
inside the same `div`.  If the anchor is pointing to a different
server, then pop up a dialog warning the user that they are leaving
the Trusted Analytics docs site, and then proceed to load the page the
old fashioned way (without AJAX).

To do:

1. Fix problems above.

2. Make this page work without JavaScript (i.e., a responsive fallback
that isn't too ugly if the user has JS disabled).  This shouldn't be
too hard since the menu is already semantically structured as a proper
list with `ul` and `li` tags.